Added retry in s3 test lib.

+++ libs/s3/s3_test_lib.py
@@ -32,7 +32,9 @@ from botocore.client import Config
 from commons import commands
 from commons import errorcodes as err
 from commons.exceptions import CTException
-from commons.utils.system_utils import create_file, run_local_cmd
+from commons.utils.system_utils import create_file
+from commons.utils.system_utils import run_local_cmd
+from commons.utils.s3_utils import poll
 from libs.s3 import S3_CFG, ACCESS_KEY, SECRET_KEY
 from libs.s3.s3_core_lib import S3Lib
 from libs.s3.s3_acl_test_lib import S3AclTestLib
@@ -359,7 +361,10 @@ class S3TestLib(S3Lib):
         try:
             LOGGER.info("You have opted to delete buckets.")
             start_time = perf_counter()
-            response = super().delete_bucket(bucket_name, force)
+            if force:
+                response = poll(super().delete_bucket, bucket_name, force)
+            else:
+                response = super().delete_bucket(bucket_name, force)
             end_time = perf_counter()
             LOGGER.debug(response)
             LOGGER.info(
